Factor out FilterControl
