+++ src/app/common/components/FilterToolbar/FilterToolbar.tsx
@@ -1,11 +1,6 @@
 import React, { useState } from 'react';
 import {
   SelectOptionProps,
-  Select,
-  SelectOption,
-  InputGroup,
-  Button,
-  ButtonVariant,
   DataToolbar,
   DataToolbarContent,
   DataToolbarToggleGroup,
@@ -15,14 +10,16 @@ import {
   DropdownItem,
   DataToolbarFilter,
 } from '@patternfly/react-core';
-import { FilterIcon, SearchIcon } from '@patternfly/react-icons';
+import { FilterIcon } from '@patternfly/react-icons';
+
+import { FilterControl } from './FilterControl';
 
 export enum FilterType {
   select = 'select',
   search = 'search',
 }
 
-type FilterValue = any; // TODO.. SelectProps.selections? string? array?
+export type FilterValue = any; // TODO.. SelectProps.selections? string? array?
 
 export interface IFilterCategory {
   key: string;
@@ -32,87 +29,17 @@ export interface IFilterCategory {
   placeholderText?: string; // TODO only if select type?
 }
 
-interface IFilterValues {
+export interface IFilterValues {
   [categoryKey: string]: FilterValue;
 }
 
-interface IFilterControlProps {
-  category: IFilterCategory;
-  value: FilterValue;
-}
-
-interface IFilterToolbarProps {
+export interface IFilterToolbarProps {
   filterCategories: IFilterCategory[];
   filterValues: any[][]; // TODO type this...
   setFilterValues: (values: IFilterValues) => void;
 }
 
-const FilterControl: React.FunctionComponent<IFilterControlProps> = ({ category, value }) => {
-  const [isFilterDropdownOpen, setIsFilterDropdownOpen] = useState(false);
-  // @ts-ignore
-  const onFilterSelect = () => console.log('onFilterSelect', arguments);
-
-  // TODO split into sub components?
-
-  const [inputValue, setInputValue] = useState('');
-
-  const onInputSubmit = () => console.log('TODO');
-
-  // @ts-ignore
-  const onInputChange = () => console.log('onInputChange', arguments);
-
-  const onInputKeyDown = () => {
-    onInputSubmit();
-    // @ts-ignore
-    console.log('onInputKeyDown', arguments);
-  };
-
-  if (category.type === FilterType.select) {
-    return (
-      <Select
-        aria-label={category.title}
-        onToggle={() => setIsFilterDropdownOpen(!isFilterDropdownOpen)}
-        onSelect={onFilterSelect} // TODO ???
-        selections={value}
-        isExpanded={isFilterDropdownOpen}
-        placeholderText="Any"
-      >
-        {category.selectOptions.map(optionProps => (
-          <SelectOption {...optionProps} />
-        ))}
-      </Select>
-    );
-  }
-  if (category.type === FilterType.search) {
-    const id = `${category.key}-input`;
-    return (
-      <InputGroup>
-        {/*
-          // @ts-ignore issue: https://github.com/konveyor/mig-ui/issues/747 */}
-        <TextInput
-          name={id}
-          id={id}
-          type="search"
-          aria-label={`${category.title} filter`}
-          onChange={onInputChange}
-          value={inputValue}
-          placeholder={category.placeholderText}
-          onKeyDown={onInputKeyDown}
-        />
-        <Button
-          variant={ButtonVariant.control}
-          aria-label="search button for search input"
-          onClick={onInputSubmit}
-        >
-          <SearchIcon />
-        </Button>
-      </InputGroup>
-    );
-  }
-  return null;
-};
-
-const FilterToolbar: React.FunctionComponent<IFilterToolbarProps> = ({
+export const FilterToolbar: React.FunctionComponent<IFilterToolbarProps> = ({
   filterCategories,
   filterValues,
   setFilterValues,
@@ -127,7 +54,9 @@ const FilterToolbar: React.FunctionComponent<IFilterToolbarProps> = ({
 
   const clearAllFilters = () => console.log('clearAllFilters!'); // TODO
   const onClearFilter = () => console.log('onClearFilter'); // TODO
-  const onFilterChange = () => console.log('onFilterChange'); // TODO
+
+  const setFilterValue = (category: IFilterCategory, newValue: FilterValue) =>
+    setFilterValues({ ...filterValues, [category.key]: newValue });
 
   const currentFilterCategory = filterCategories.find(
     category => category.key === currentCategoryKey
@@ -146,7 +75,6 @@ const FilterToolbar: React.FunctionComponent<IFilterToolbarProps> = ({
               }
               isOpen={isCategoryDropdownOpen}
               dropdownItems={filterCategories.map(category => (
-                // TODO properties for row property name and title
                 <DropdownItem key={category.key} onClick={() => onCategorySelect(category)}>
                   {category.title}
                 </DropdownItem>
@@ -157,11 +85,15 @@ const FilterToolbar: React.FunctionComponent<IFilterToolbarProps> = ({
             <DataToolbarFilter
               key={category.key}
               chips={filterValues[category.key]}
-              deleteChip={onClearFilter}
+              deleteChip={onClearFilter} // TODO is this one value or all of a category?
               categoryName={category.title}
               showToolbarItem={currentFilterCategory.key === category.key}
             >
-              <FilterControl category={category} value={filterValues[category.key]} />
+              <FilterControl
+                category={category}
+                value={filterValues[category.key]}
+                setValue={newValue => setFilterValue(category, newValue)}
+              />
             </DataToolbarFilter>
           ))}
         </DataToolbarToggleGroup>
@@ -169,5 +101,3 @@ const FilterToolbar: React.FunctionComponent<IFilterToolbarProps> = ({
     </DataToolbar>
   );
 };
-
-export default FilterToolbar;
